+++ huginn-proxy-lib/tests/config/loader.rs
@@ -1,16 +1,8 @@
 use std::fs;
-use std::path::PathBuf;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use huginn_proxy_lib::config::load_from_path;
 
-fn tmp_path(name: &str) -> PathBuf {
-    let nanos = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_else(|_| Duration::from_secs(0))
-        .as_nanos();
-    std::env::temp_dir().join(format!("huginn-{nanos}-{name}.toml"))
-}
+use super::tmp_path;
 
 #[test]
 fn loads_minimal_config() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
@@ -284,200 +276,6 @@ policy = "default-src 'none'"
         .ok_or("route headers should be present")?;
     assert!(headers.csp.enabled);
     assert_eq!(headers.csp.policy.expose(), "default-src 'none'");
-
-    let _ = fs::remove_file(&path);
-    Ok(())
-}
-
-#[test]
-fn audit_warns_when_domain_override_disables_enabled_global_rate_limit(
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    use huginn_proxy_lib::config::security_override_warnings;
-
-    let path = tmp_path("audit-domain-rl");
-    let toml = r#"
-listen = { addrs = ["127.0.0.1:0"] }
-backends = [{ address = "backend:9000" }]
-
-[security.rate_limit]
-enabled = true
-requests_per_second = 100
-
-[[domains]]
-host = "api.example.com"
-routes = [{ prefix = "/", backend = "backend:9000" }]
-
-[domains.security.rate_limit]
-enabled = false
-"#;
-    fs::write(&path, toml)?;
-    let cfg = load_from_path(&path)?;
-
-    let warnings = security_override_warnings(&cfg);
-    assert_eq!(warnings.len(), 1, "expected one finding, got: {warnings:?}");
-    assert_eq!(warnings[0].scope, "domain 'api.example.com'");
-    assert!(warnings[0].message.contains("rate_limit"));
-
-    let _ = fs::remove_file(&path);
-    Ok(())
-}
-
-#[test]
-fn audit_warns_when_route_headers_override_drops_global_csp(
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    use huginn_proxy_lib::config::security_override_warnings;
-
-    let path = tmp_path("audit-route-csp");
-    let toml = r#"
-listen = { addrs = ["127.0.0.1:0"] }
-backends = [{ address = "backend:9000" }]
-
-[security.headers.csp]
-enabled = true
-policy = "default-src 'self'"
-
-[[domains]]
-host = "api.example.com"
-
-[[domains.routes]]
-prefix = "/admin"
-backend = "backend:9000"
-
-# Route sets a headers block that does NOT re-enable CSP → CSP dropped (whole-block replace).
-[domains.routes.security.headers.hsts]
-enabled = true
-"#;
-    fs::write(&path, toml)?;
-    let cfg = load_from_path(&path)?;
-
-    let warnings = security_override_warnings(&cfg);
-    assert_eq!(warnings.len(), 1, "expected one finding, got: {warnings:?}");
-    assert!(warnings[0].scope.starts_with("route '/admin'"));
-    assert!(warnings[0].message.contains("CSP"));
-
-    let _ = fs::remove_file(&path);
-    Ok(())
-}
-
-#[test]
-fn audit_silent_when_override_tightens_or_parent_inactive(
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    use huginn_proxy_lib::config::security_override_warnings;
-
-    let path = tmp_path("audit-silent");
-    // Global rate limit disabled; domain ENABLES it (tightening) → no footgun.
-    let toml = r#"
-listen = { addrs = ["127.0.0.1:0"] }
-backends = [{ address = "backend:9000" }]
-
-[[domains]]
-host = "api.example.com"
-routes = [{ prefix = "/", backend = "backend:9000" }]
-
-[domains.security.rate_limit]
-enabled = true
-requests_per_second = 5
-"#;
-    fs::write(&path, toml)?;
-    let cfg = load_from_path(&path)?;
-
-    assert!(security_override_warnings(&cfg).is_empty());
-
-    let _ = fs::remove_file(&path);
-    Ok(())
-}
-
-#[test]
-fn header_audit_warns_on_duplicate_add_and_add_remove_conflict(
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    use huginn_proxy_lib::config::header_config_warnings;
-
-    let path = tmp_path("hdr-dup");
-    let toml = r#"
-listen = { addrs = ["127.0.0.1:0"] }
-backends = [{ address = "backend:9000" }]
-
-[headers.request]
-# X-Foo added twice (last wins) and also removed (contradictory).
-add = [
-  { name = "X-Foo", value = "a" },
-  { name = "x-foo", value = "b" },
-]
-remove = ["X-Foo"]
-
-[headers.response]
-add = [{ name = "X-Bar", value = "1" }]
-"#;
-    fs::write(&path, toml)?;
-    let cfg = load_from_path(&path)?;
-
-    let warnings = header_config_warnings(&cfg);
-    assert_eq!(warnings.len(), 2, "expected dup + conflict, got: {warnings:?}");
-    assert!(warnings.iter().all(|w| w.scope == "global headers"));
-    assert!(warnings
-        .iter()
-        .any(|w| w.message.contains("added more than once")));
-    assert!(warnings
-        .iter()
-        .any(|w| w.message.contains("both added and removed")));
-
-    let _ = fs::remove_file(&path);
-    Ok(())
-}
-
-#[test]
-fn header_audit_warns_on_duplicate_custom_security_header(
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    use huginn_proxy_lib::config::header_config_warnings;
-
-    let path = tmp_path("hdr-custom-dup");
-    let toml = r#"
-listen = { addrs = ["127.0.0.1:0"] }
-backends = [{ address = "backend:9000" }]
-
-[security.headers]
-custom = [
-  { name = "X-Bar", value = "1" },
-  { name = "X-BAR", value = "2" },
-]
-"#;
-    fs::write(&path, toml)?;
-    let cfg = load_from_path(&path)?;
-
-    let warnings = header_config_warnings(&cfg);
-    assert_eq!(warnings.len(), 1, "expected one finding, got: {warnings:?}");
-    assert_eq!(warnings[0].scope, "global security headers");
-    assert!(warnings[0].message.contains("listed more than once"));
-
-    let _ = fs::remove_file(&path);
-    Ok(())
-}
-
-#[test]
-fn header_audit_silent_on_cross_scope_override(
-) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    use huginn_proxy_lib::config::header_config_warnings;
-
-    let path = tmp_path("hdr-cross-scope");
-    // Same header name at global and domain scope is an intentional override, not a duplicate.
-    let toml = r#"
-listen = { addrs = ["127.0.0.1:0"] }
-backends = [{ address = "backend:9000" }]
-
-[headers.request]
-add = [{ name = "X-Foo", value = "global" }]
-
-[[domains]]
-host = "api.example.com"
-routes = [{ prefix = "/", backend = "backend:9000" }]
-
-[domains.headers.request]
-add = [{ name = "X-Foo", value = "domain" }]
-"#;
-    fs::write(&path, toml)?;
-    let cfg = load_from_path(&path)?;
-
-    assert!(header_config_warnings(&cfg).is_empty());
 
     let _ = fs::remove_file(&path);
     Ok(())
